port to multi-target with `netstandard2.1` and `net8.0`
Godot wants net 8, not 10
should enable this to work with Unity, too

+++ SimLynx/Core/Phasing/Phase.cs
@@ -10,7 +10,7 @@ namespace SimLynx.Core.Phasing;
 /// </summary>
 public abstract class Phase(IIndex<Symbol, IPhaseManager> phaseManagers) : IPhase
 {
-    private byte _startFlag = 0;
+    private int _startFlag = 0;
 
     /// <inheritdoc/>
     public bool HasStarted => _startFlag != 0;
